% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-us-stock-prices.R
\name{all_us_stock_prices}
\alias{all_us_stock_prices}
\title{Get all US Stock Prices}
\usage{
all_us_stock_prices()
}
\value{
Dataframe with last price, volume for the curernt day, and percent change
since open.
}
\description{
Show a metric plotted over time as a line chart.
}
\examples{
\dontrun{
all_us_stock_prices(metric_df)
}

}
